Move all external typings to metacom.d.ts
