--- lib/websocket/ws.d.ts
@@ -1,7 +1,0 @@
-export { OPCODES, CLOSE_CODES } from './constants.js';
-export { WebsocketServer, MAGIC } from './server.js';
-export type { WebsocketServerOptions, VerifyClientInfo } from './server.js';
-export { Connection, CLOSE_TIMEOUT } from './connection.js';
-export type { ConnectionOptions } from './connection.js';
-export { Frame } from './frame.js';
-export { FrameParser, ParseError, PARSE_ERR_CODES } from './frameParser.js';
